resolve controller cannot talk to consumer in fuel

+++ tests/run.sh
@@ -114,8 +114,15 @@ prepare_compute_ssh() {
 }
 
 get_consumer_ip() {
-    CONSUMER_IP=$(sudo ssh $ssh_opts root@$INSTALLER_IP \
-                  "ip route get $COMPUTE_IP | awk '/ src /{print \$NF}'")
+    if [[ "$INSTALLER_TYPE" == "apex" ]] ; then
+        CONSUMER_IP=$(sudo ssh $ssh_opts root@$INSTALLER_IP \
+                      "ip route get $COMPUTE_IP | awk '/ src /{print \$NF}'")
+    elif [[ "$INSTALLER_TYPE" == "fuel" ]] ; then
+        CONSUMER_IP=$(sudo sshpass -p r00tme ssh $ssh_opts root@${INSTALLER_IP} \
+                      "ip route get $COMPUTE_IP | awk '/ src /{print \$NF}'")
+    elif [[ "$INSTALLER_TYPE" == "local" ]] ; then
+        CONSUMER_IP=$(ip route get $COMPUTE_IP | awk '/ src /{print $NF}')
+    fi
     echo "CONSUMER_IP=$CONSUMER_IP"
 
     if [[ -z "$CONSUMER_IP" ]]; then
@@ -211,16 +218,23 @@ start_consumer() {
     # NOTE(r-mibu): create tunnel to the installer node, so that we can
     # avoid some network problems dpends on infra and installers.
     # This tunnel will be terminated by stop_consumer() or after 10 mins passed.
-    TUNNEL_COMMAND="sudo ssh $ssh_opts $INSTALLER_IP -R $CONSUMER_PORT:localhost:$CONSUMER_PORT 'sleep 600'"
-    $TUNNEL_COMMAND > ssh_tunnel.log 2>&1 < /dev/null &
+    if [[ "$INSTALLER_TYPE" == "apex" ]] ; then
+        TUNNEL_COMMAND="sudo ssh $ssh_opts $INSTALLER_IP -R $CONSUMER_PORT:localhost:$CONSUMER_PORT 'sleep 600'"
+        $TUNNEL_COMMAND > ssh_tunnel.log 2>&1 < /dev/null &
+    elif [[ "$INSTALLER_TYPE" == "fuel" ]] ; then
+        TUNNEL_COMMAND="sudo sshpass -p r00tme ssh $ssh_opts $INSTALLER_IP -R $CONSUMER_PORT:localhost:$CONSUMER_PORT 'sleep 600'"
+        $TUNNEL_COMMAND > ssh_tunnel.log 2>&1 < /dev/null &
+    fi
 }
 
 stop_consumer() {
     pgrep -f "python consumer.py" || return 0
     kill $(pgrep -f "python consumer.py")
     print_log consumer.log
-    kill $(pgrep -f "$TUNNEL_COMMAND")
-    print_log ssh_tunnel.log
+    if [[ "$INSTALLER_TYPE" -ne "local" ]] ; then
+        kill $(pgrep -f "$TUNNEL_COMMAND")
+        print_log ssh_tunnel.log
+    fi
 }
 
 wait_for_vm_launch() {
